+++ crates/core/src/config.rs
@@ -247,6 +247,9 @@ pub struct ProtoSettingsConfig {
     #[setting(default = default_builtin_plugins)]
     pub builtin_plugins: BuiltinPlugins,
 
+    #[setting(env = "PROTO_CACHE_DURATION")]
+    pub cache_duration: Option<u64>,
+
     #[setting(env = "PROTO_DETECT_STRATEGY")]
     pub detect_strategy: DetectStrategy,
 
@@ -468,7 +471,7 @@ impl ProtoConfig {
             self.plugins.insert(
                 Id::raw(SCHEMA_PLUGIN_KEY),
                 PluginLocator::Url(Box::new(UrlLocator {
-                    url: "https://github.com/moonrepo/plugins/releases/download/schema_tool-v0.17.3/schema_tool.wasm".into()
+                    url: "https://github.com/moonrepo/plugins/releases/download/schema_tool-v0.17.4/schema_tool.wasm".into()
                 }))
             );
         }